Add an initial version of a Resolved CFD-DEM doc (#1831)
Description
We do not have any theory guide on the documentation for the Resolved CFD-DEM. Considering the fact that the paper for Lethe 1.0 has been released a while ago, this can be remedied. This integrates some of the Resolved CFD-DEM equations and documentation from the 1.0 release paper inside of the theory guide. It is maybe not perfect, but I think this is much better than having nothing.


Co-authored-by: Olivier Guévremont <guevremont.o@gmail.com>