Account for ALE case in velocity for stabilization (#1834)
Description
The velocity used to compute the stabilization parameter did not account for the ALE velocity component when the mortar feature is enabled. This led to inconsistent stabilization terms in the rotor and stator sides, and it has been fixed in this PR.

Solution
The velocity used to compute the parameter tau now takes into account velocity_ale if mortar is enabled. This is fixed in both matrix-free and matrix-based solvers.

Testing
No new tests have been added. Eigenvalues of the multigrid matrix-free TGV example had small changes.