Add option to fix pressure node in matrix-based solver (#1828)
Description
This PR adds the option to fix pressure in a node within the matrix-based solver. The process is the same as what is already done in the coarse multigrid level of the matrix-free solver when fix pressure constant is set to true.

Testing
One existent test (mortar_taylorcouette_transient) was modified to have fix pressure constant set to true. This led to very small changes in the torque results.

Documentation
No new parameters are needed, since fix pressure constant already exists.