Make reaction model more robust to low  reaction orders (#1825)
Description
The reaction model included in the tracer worked as intended for most cases, but failed when the reaction order was set below 1 due to the computation of a negative power of concentration (when =0).

Solution
This PR solves this issue by using instead C_effective = (C^2 + epsilon^2)^0.5, which ensures:

Smooth transition from C to epsilon (when it would be 0)
Only positive values
Limited stiffness in the linear solver of the tracer physics
A parameter to set explicitly epsilon is added to control this behavior precisely.

Testing
All tests still pass.

A new test at order=0.5 and null initial condition, which previously caused a numerical error, is added. This test uses a MMS case to verify that the proper solution is recovered.

Documentation
The new parameter is added to the documentation.



Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>
Co-authored-by: Bruno Blais <blais.bruno@gmail.com>