Fix the VANS example of a flow through a packed bed by using an implicit drag formulation instead of the default semi-implicit one. (#1840)
Description
Before the 1.0.3 release, I have realized that the flow through a packed bed example was not very well set-up. It was solving the flow in steady-state, but using the default semi-implicit drag formulation leading to an incorrect pressure drop. Furthermore, our default robust model is really model A and not model B. Finally, we never really ouput the pressure drop in the terminal while this is an easy thing to do.

Solution
I did the following things

Switch the formulation to model A
Use implicit drag formulation
Update and refresh the documentation
Testing
The test now runs and give adequate results.

Documentation
Documentation for the case has also been updated.

Co-authored-by: antonellasuccar <antonella.succar@polymtl.ca>