Remove interdependency of core on DEM (#1842)
Description
The SerialSolid class, which is in the core library, had become dependent on the DEM library which is not something that fits our architecture. This was because of a single functor used for comparisons. This has been fixed by moving this functions to utilities.h.

Testing
All test pass

Documentation
Nothing to do here.