Simplify some elements of the documentation (#1843)
Description
The documentation of the SimulationControl and the Analytical solution sections were a bit convoluted. I have tried to simplify them.

Futhermore, some parameter indentation were not correctly done inside of the .rst file which led to boxes that were slightly out of alignment (not at the right indentation level). I have fixed that also.

Co-authored-by: Olivier Guévremont <guevremont.o@gmail.com>