Improve extract slice tool (#1848)
Description
The extract-slice-from-vtu.py tool was not working properly with --np > 1 because an argument (file path) was missing in the call of extract_slice.

When the flag group files in the .prm was set to >1, the extract-slice-from-vtu.py would not slice all .vtus associated with a .pvtu.

There was no sliced .pvd file outputted, so the slices could not be used easily in other post-processing tools using pvd.

Solution
The missing argument was added.

pyvista can read pvtus correctly, so the list of files to read is based on pvtus instead of vtus.

The pvd file of the slices is now outputted.

Testing
No new test added, it works now in parallel and with pvtus.

Documentation
No documentation related.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Co-authored-by: Olivier Guévremont <guevremont.o@gmail.com>