Fix while loop error on contact detection (#1855)
Description
The new implementation of the solid-particle contact detection could lead to an infinite loop for one of the cases that happens only on deformed mesh.

Solution
The issue was that two continue statement could be called without having the iterator increased. This lead to the aforementioned bug. Running any example with a solid object (bunny drill or granular mixer) would directly lead to this infinite loop being encountered no matter the number of cores. The solution is simple (increment the iterator), but it also demonstrates that we need more robust testing for this part of the code.

Testing
This has been tested by running both the Bunny Drill and the granular mixer which now both work. @OGaboriault I will leave to you to do a follow-up PR where you add an application_test with a non-trivial moving geometry. I think something like the granular mixer, but with less particles, could be a quite adequate application_test for this.

Documentation
Nothing to document here.