Dem distribution update tests (#1849)
Description
Follow up to PR 1837. This PR adds a message to the user at the beginning of the simulation. This message informs the user about the distribution type used by every particle type. If a normal or lognormal distribution is used and if the cutoff value are cutting off more than 25% of the distribution weighted on number, a warning is shown.

Testing
Since this PR adds a message at the start of a DEM simulation, every test need to be updated. This is why this PR is done separately from the lognormal distribution PR.

Documentation
I did some minor changes to the update-golden documentation since it wasn't clear in my option.