Fix nomenclature for forces in the unresolved
The nomenclature we use for the forces in the unresolved CFD-DEM
had some issues and incoherence. Finally I think it is better to use the
lower case nomenclature for the volumetric source term and the capital
case nomenclature for the forces on the particles following the last
preprint by O Gaboriault.