Electromagnetic physical properties (#1862)
Description
This PR add the material properties that will be used by the time-harmonic electromagnetic solver which include the electric conductivity, the complex electric permittivity and the complex magnetic permeability. At the moment, those properties only support a "constant" field and the dimensionality have not been updated as it is common to use already relative properties (i.e. normalized by the void properties). A unit test for the electromagnetic constant properties was added, along with the update of the already existing "physical_properties_manager" tests.

Testing
All new and updated tests pass.

Documentation
The documentation will come in a future PR.

"electric conductivity": The inverse of the resistivity of a material to an applied current.
"electric permittivity real": The ability of a material to store electromagnetic energy by polarization.
"electric permittivity real": The ability of a material to dissipate electromagnetic energy by polarization.
"magnetic permeability real": The ability of a material to store electromagnetic energy by magnetization.
"magnetic permeability imag": The ability of a material to dissipate electromagnetic energy by magnetization.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>