Fix cfd dem matrix free restart bug in parallel due to wrong vector size. (#1863)
Description
The unresolved CFD-DEM matrix-free solver would crash when restarting simulations with more than 1 mpi rank.

Solution
Deal.ii parallel distributed vectors require that the locally_relevant component of the vector be initialized before reading a vector checkpoint. This was not the case for the trilinos vectors. Consequently, the read of the velocity vector would crash when restarted in parallel

Testing
It will be a bit annoying to do a unit test for this. However I have tested on my machine and this work. This is something that will be added to the validation test later.

Documentation
Nothing to document here.