Fix the time history of the void fraction when restarting matrix-free CFD-DEM simulations (#1864)
Description and Solution
This is a follow-up of #1863 . Although #1863 fixed the restart and prevented the simulations from crashing, the time history of the void fraction was not store appropriately. This stemmed from a confusion since in the matrix-free solver, it is the deal.II distributed vectors for the void fraction which are checkpointed and not the Trilinos ones. This PR reads the correct deal.II vector when reading a checkpoint, but also ensures that the values in the Trilinos vectors matches that of the deal.II vectors. This allows reproducing the norm of the residuals to machine accuracy when restarting. This fixes restarting with everything enabled. I will come up with a follow-up PR soon with a nice application test for this.

Testing
Testing on my machine and on Trillium

Documentation
Nothing to document