Dem floating wall declaration (#1850)
Description
This PR fixes issue #1360
The floating wall declaration was using 6 different parameters to define the point on wall and the normal vector. Now, these two thing are defined using two parameters.

Also found a bug relative to the floating walls. The normal vector was never normalized, which is messing with the normal overlap (probably). One of our test was defining the normal vector as a non unit vector, which is fine by itself (the user should be able to do that). This test was falling after my change, which is why the output needed to be updated. Other tests that were using floating walls were declaring the normal vector a unit vector, thus those test didn't change.

Testing
The falling test got updated.

Documentation
Documentation got updated.