Mortar refactoring in read_mesh function (#1854)
Description
The verification of the number of cells at the stator and rotor interfaces was being done inside the loops for dealii and gmsh grids. This PR adjusts this by doing the verification only once, after the initial mesh refinement is performed.

Testing
No tests were added, and nothing should change in the existing ones.