Clarify notation unresolved cfd dem (#1856)
Description
The documentation for the forces in the unresolved CFD-DEM was still unclear (and I am the one to blame for this). The nomenclature we use for the forces in the unresolved CFD-DEM had some issues and incoherence. Finally I think it is better to use the lower case nomenclature for the volumetric source term and the capital case nomenclature for the forces on the particles following the last preprint by O Gaboriault.

This PR uses this approach. Furthermore, the fact that the forces applied on the fluid are clearly (-) the forces applied on the particles is now a LOT more explicit and this is much better I believe.

Reading this also made me realize that some of the nomenclature in the DEM was still missing some mathrm there and there. I will fix this in a follow-up PR after this :).

Co-authored-by: Antonella Succar <142912677+antonellasuccar@users.noreply.github.com>
Co-authored-by: OGaboriault <102625061+OGaboriault@users.noreply.github.com>