Add support for boundary list in multiphysics boundary conditions (#1867)
Description
This is a copy from #1866 which was closed and could not be reopened properly.

This PR fixes boundary list that were added in #1697. Boundary lists were not being parsed properly for physics other than fluid dynamic. The parameter parsing expected an integer instead of a list of integers. The treatment of boundary id list within the source code was already done properly, so nothing changed.

Solution
The parsing now expected a list of integers for multiphysic boundary conditions.

Testing
I tested all physics with lists of boundary ids and they work. No more tests necessary.

Documentation
The documentation for multiphysic boundary conditions has been updated to explicitly show the availability of boundary id list feature for all physics.