Dem solid surface (#1865)
Description
This is a follow up to PR #1855, where BB requested a test to be added regarding solid-surface. This PR adds an applications test. It is a simple particle that is touching two triangles that are edge sharing. One contact is on an edge of the first triangle while the second contact is on a vertex of the second triangle.

This application test is stalling when using the commit b29f663 (the commit before the fix of PR #1855).

I also found a logical error in an if statement in the beginning of the while loop where we didn't check if the two triangles were vertex and edge sharing neighbors. I think this error is the culprit of the problem I am still having, but I am still not a hundred percent sure. (So far everything works on the clusters using this branch)

Edit :
While doing this PR, I realized that we were looping three times on the particles potentially in contact for a given triangle. The first one was to create the containers for the find_particle_triangle_projection function, the second one was inside the find_particle_triangle_projection function and the last one was to compute the overlap if the distance_check is set to true. Thus, I have decided to rework the design of the function (I was telling myself that the bug that I'm still experiencing could be fix by this change) . Following this refactoring, I have observed a speed up from 895 seconds on master branch to 345 seconds using this branch when running the bunny drill example (only the drilling part).

Solution
Testing
A new application test is added to lethe. Every old test still passes even after the refactoring.